+++ crates/ruma-common/tests/serde/mod.rs
@@ -1,4 +1,2 @@
 mod empty_strings;
 mod enum_derive;
-mod url_deserialize;
-mod url_serialize;