+++ crates/ruma-common/CHANGELOG.md
@@ -43,6 +43,8 @@ Breaking changes:
   now removed)
 * Move the `Unsigned` associated type from `StateEventContent` to `OriginalStateEventContent`
   * `Redacted*EventContent`s don't have an `unsigned` type anymore
+* Remove the `serde::urlencoded` module
+  * Query string (de)serialization is now done by the `serde_html_form` crate
 
 Improvements: